refact: ColorBox 컴포넌트 분리
